validation

+++ src/Adminpanel/Components/Header/SidenavData.jsx
@@ -6,7 +6,7 @@ import {
   ListItemIcon,
   ListItemText,
 } from "@material-ui/core";
-import { NavLink } from "react-router-dom";
+import { NavLink} from "react-router-dom";
 import { useStyles } from "./HeaderStyle";
 
 import DashboardIcon from "@material-ui/icons/Dashboard";
@@ -14,12 +14,63 @@ import UploadFileIcon from '@mui/icons-material/UploadFile';
 import PostAddIcon from "@material-ui/icons/PostAdd";
 import AnalyticsIcon from '@mui/icons-material/Analytics';
 import ExitToAppIcon from "@material-ui/icons/ExitToApp";
+import axios from "axios";
 
 export default function SidenavData({ handleDrawerClose }) {
 
   const adminLogout = () => {
-    localStorage.clear();
-    window.location.href = "/adminlogin";
+    // localStorage.clear();
+    // window.location.href = "/adminlogin";
+    const token = localStorage.getItem("logintoken")
+    console.log(token)
+    const hello = (token) => {
+      axios(
+        {
+          url: "http://localhost:5001/api/admin/logOut",
+          method: "delete",
+          headers: {
+                  "Authorization": `Bearer ${token}`
+                }
+        }
+      )
+        .then((response) => {
+          console.log(response);
+          if (response.status === 200) {
+            // const logintoken = response.data.token
+            // console.log(response.data.token)
+            // localStorage.setItem('logintoken', logintoken)
+              // window.location.href="/dashboard";
+                        localStorage.clear();
+                         window.location.href = "/adminlogin";
+                        // setTimeout(()=>{
+                        //   history.push('/adminlogin');
+
+                        // },1000)
+          }
+        })
+        .catch((err) => {
+          console.log(err);
+        })
+    }
+    hello(token)
+    // const api = axios(
+    //   {
+    //     url: "http://localhost:5001/api/admin/logOut",
+    //     method: "DELETE",
+    //     headers: {
+    //       "Authorization": `Bearer ${token}`
+    //     }
+    //   }.then((Response) => {
+    //     console.log(Response)
+    //     // if (Response.status === 200) {
+    //     //   localStorage.clear();
+    //     //   // window.location.href = "/adminlogin";
+    //     // }
+    //   }).catch((err) => {
+    //     console.log(err);
+    //   })
+    // )
+    console.log(hello)
   }
     // let res = axios.delete("http://localhost:5001/api/admin/logOut");
     // if(res.status===204)
@@ -28,22 +79,7 @@ export default function SidenavData({ handleDrawerClose }) {
     //   window.location.href="/adminlogin"; 
     //   console.log("success")
     // }
-    // axios(
-    //   {
-    //     url: "http://localhost:5001/api/admin/logOut",
-    //     method: "delete",
-    //     headers: {
-    //       "Authorization": `Bearer ${token}`
-    //     }
-    //   }.then((Response) => {
-    //     if (Response.status === 200) {
-    //       localStorage.clear();
-    //       window.location.href = "/adminlogin";
-    //     }
-    //   }).catch((err) => {
-    //     console.log(err);
-    //   })
-    // )
+   
   //   const headers = { 
   //     'Authorization': 'Bearer my-token',
   //     'My-Custom-Header': 'foobar'
@@ -69,28 +105,37 @@ export default function SidenavData({ handleDrawerClose }) {
     },
     { label: "Admin Details", link: "/admindetails", icon: <ExitToAppIcon /> },
     { label: "Blog", link: "/blog", icon: <ExitToAppIcon /> },
-    { label: "Logout", link: "/", icon: <ExitToAppIcon onClick={adminLogout} /> },
+    // { label: "Logout", link : "/",  icon: <ExitToAppIcon onClick={adminLogout} /> },
   ];
 
   return (
+    <>
     <List>
-      {listItemData.map((item, i) => (
+      {listItemData.map((item) => (
         <Button
           size='small'
           onClick={() => handleDrawerClose()}
           className={classes.navButton}>
           <ListItem
             exact
-            key={i}
+            key="ListItem"
             component={NavLink}
-            to={item.link}
+            to={item.link && item.link}
             className={classes.navlink}
             activeClassName={classes.selectedNav}>
             <ListItemIcon>{item.icon}</ListItemIcon>
             <ListItemText sx={{ mb: 5 }}>{item.label}</ListItemText>
           </ListItem>
         </Button>
+
       ))}
+      {/* <button onClick={adminLogout}>
+        Logout
+      </button> */}
     </List>
+     <Button size="small" onClick={adminLogout}>
+     Logout
+   </Button>
+   </>
   );
 }
